requirements.txt
asgiref==3.4.1
attrs==21.2.0
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
graphviz==0.17
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
